Snakemake workflow now uses separate compile rules for all vs final test
Test files can be inputted as wildcards.
I avoid overwriting of the GRNfile by setting it as a resource, so it 
must be set as GRNfile=1 during snakemake call
NOTE: current version checks that ALL problems are trained. This is a 
feature. The problem set would be incomplete if only a few were trained.