Do not compile before starting a new simulation
Simulations read the GRNfile when the program starts, so if we compile all first and then start only the GRNs in the final setup will be run.
I solve this by making the setup rules wait for 10 seconds, so that the simulations are launched before the new compilation starts.
Setting lower priority ensures that when cores are limited the runs always work before the setup of another rule.
NOTE: this only works if we throttle simultaneous execution of setup rules via the GRNfile=1 resource